app/database.py
import os
import psycopg2

def get_db_connection():
    """Establishes a connection to the PostgreSQL database."""
    try:
        conn = psycopg2.connect(os.environ.get('DATABASE_URL'))
        return conn
    except psycopg2.OperationalError as e:
        print(f"DATABASE CONNECTION ERROR: {e}")
        return None
